Move UrlInputFragment to different package
We are going to add more related class to UrlInput.
